build: automate release step fully
